// frontend/src/generated/api/models/ImportProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ImportProfile = {
    Uci: 'uci',
    GenericCsv: 'generic_csv'
} as const;
export type ImportProfile = typeof ImportProfile[keyof typeof ImportProfile];


export function instanceOfImportProfile(value: any): boolean {
    for (const key in ImportProfile) {
        if (Object.prototype.hasOwnProperty.call(ImportProfile, key)) {
            if (ImportProfile[key as keyof typeof ImportProfile] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ImportProfileFromJSON(json: any): ImportProfile {
    return ImportProfileFromJSONTyped(json, false);
}

export function ImportProfileFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImportProfile {
    return json as ImportProfile;
}

export function ImportProfileToJSON(value?: ImportProfile | null): any {
    return value as any;
}

export function ImportProfileToJSONTyped(value: any, ignoreDiscriminator: boolean): ImportProfile {
    return value as ImportProfile;
}